libraries/WiFi: Renamed Socket for Win case insensitive fix.
